refactor: modularize renderer
