+++ index.html
@@ -111,6 +111,6 @@
          </aside>
     </div>
 
-    <script defer src="renderer.js"></script>
+    <script type="module" src="renderer.js"></script>
 </body>
 </html>
